# Support general 2D covariance in reprojection error. Refactor cost functions