Relevant: https://github.com/colmap/colmap/pull/2783

This PR includes the updates as follows:
1. A more unified interface for applying covariance in cost function, with templating and constexpr. Support three types of covariance with different computations: identity, diagonal, and general.
2. Apply covariance optionally in bundle adjustment. Performance and interfaces remain unchanged with identity covariance.
3. Deprecate IsotropicNoiseReprojError as it is a subset of what can be achieved with diagonal covariance with same performance. Also actually, I believe IsotropicNoiseReprojError can be achieved by ScaledLoss directly, while diagonal covariance supports resizing without ratio kept. 
4. Pybind support. Keep the classic reprojection error interface similar to before. However, deprecate stddev and change all the non-identity covariance support to ``WeightedReprojErrorCost``. This will potentially affect some pycolmap users but adapting is very simple (and I believe necessary for the long term)
5. Also adapt other cost functions with the same convention.

Example on how to adapt the previous pycolmap factors with ``stddev`` to the new one:
```
# old: deprecated!!
# cost = pycolmap.cost_functions.ReprojErrorCost(camera_model, stddev, point2d)
# new:
cost = pycolmap.cost_functions.WeightedReprojErrorCost(camera_model, 
                                                       pycolmap.CovarianceType.DIAGONAL, 
                                                       point2d, 
                                                       np.array([[stddev, 0], [0, stddev]]))
```