@B1ueber2y 2024-09-23
Benchmarking results from https://github.com/colmap/colmap/tree/main/benchmark

```
Running ./benchmark_cost_functions
Run on (16 X 5100 MHz CPU s)
CPU Caches:
L1 Data 32 KiB (x8)
L1 Instruction 32 KiB (x8)
L2 Unified 256 KiB (x8)
L3 Unified 16384 KiB (x1)
Load Average: 0.34, 1.38, 1.34
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.

main branch
-----------------------------------------------------------------------------------------------
Benchmark                                                     Time             CPU   Iterations
-----------------------------------------------------------------------------------------------
BM_ReprojErrorCostFunction/Run_mean                       46385 ns        46384 ns           50
BM_ReprojErrorCostFunction/Run_median                     46353 ns        46353 ns           50
BM_ReprojErrorCostFunction/Run_stddev                      89.7 ns         89.4 ns           50
BM_ReprojErrorCostFunction/Run_cv                          0.19 %          0.19 %            50
BM_ReprojErrorConstantPoseCostFunction/Run_mean           36861 ns        36860 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_median         36849 ns        36849 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_stddev          83.5 ns         83.2 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_cv              0.23 %          0.23 %            50
BM_ReprojErrorConstantPoint3DCostFunction/Run_mean        40875 ns        40875 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_median      40862 ns        40862 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_stddev       61.8 ns         61.7 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_cv           0.15 %          0.15 %            50

This PR
-----------------------------------------------------------------------------------------------
Benchmark                                                     Time             CPU   Iterations
-----------------------------------------------------------------------------------------------
BM_ReprojErrorCostFunction/Run_mean                       46402 ns        46400 ns           50
BM_ReprojErrorCostFunction/Run_median                     46357 ns        46357 ns           50
BM_ReprojErrorCostFunction/Run_stddev                      99.2 ns         98.4 ns           50
BM_ReprojErrorCostFunction/Run_cv                          0.21 %          0.21 %            50
BM_ReprojErrorConstantPoseCostFunction/Run_mean           36398 ns        36397 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_median         36371 ns        36370 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_stddev          93.1 ns         92.5 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_cv              0.26 %          0.25 %            50
BM_ReprojErrorConstantPoint3DCostFunction/Run_mean        40746 ns        40745 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_median      40691 ns        40690 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_stddev        217 ns          217 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_cv           0.53 %          0.53 %            50
```