@B1ueber2y 2024-09-23
More results : )

```
This PR with CovarianceType::DIAGONAL
-----------------------------------------------------------------------------------------------
Benchmark                                                     Time             CPU   Iterations
-----------------------------------------------------------------------------------------------
BM_ReprojErrorCostFunction/Run_mean                       47397 ns        47396 ns           50
BM_ReprojErrorCostFunction/Run_median                     47352 ns        47350 ns           50
BM_ReprojErrorCostFunction/Run_stddev                       104 ns          103 ns           50
BM_ReprojErrorCostFunction/Run_cv                          0.22 %          0.22 %            50
BM_ReprojErrorConstantPoseCostFunction/Run_mean           37748 ns        37747 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_median         37724 ns        37724 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_stddev          77.5 ns         76.8 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_cv              0.21 %          0.20 %            50
BM_ReprojErrorConstantPoint3DCostFunction/Run_mean        42638 ns        42637 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_median      42602 ns        42602 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_stddev       89.1 ns         88.9 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_cv           0.21 %          0.21 %            50

This PR with CovarianceType::GENERAL
-----------------------------------------------------------------------------------------------
Benchmark                                                     Time             CPU   Iterations
-----------------------------------------------------------------------------------------------
BM_ReprojErrorCostFunction/Run_mean                       56395 ns        56394 ns           50
BM_ReprojErrorCostFunction/Run_median                     56352 ns        56352 ns           50
BM_ReprojErrorCostFunction/Run_stddev                       131 ns          131 ns           50
BM_ReprojErrorCostFunction/Run_cv                          0.23 %          0.23 %            50
BM_ReprojErrorConstantPoseCostFunction/Run_mean           44858 ns        44858 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_median         44838 ns        44838 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_stddev           109 ns          109 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_cv              0.24 %          0.24 %            50
BM_ReprojErrorConstantPoint3DCostFunction/Run_mean        50473 ns        50472 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_median      50434 ns        50433 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_stddev        116 ns          116 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_cv           0.23 %          0.23 %            50
```